requirements.txt
python-telegram-bot==13.7
SQLAlchemy==1.4.48
PyMySQL==1.0.3